+++ react/IconImage.js
@@ -1,0 +1,3 @@
+import IconImage from './components/icon/Image/index'
+
+export default IconImage