+++ react/IconItalic.js
@@ -1,0 +1,3 @@
+import IconItalic from './components/icon/Italic/index'
+
+export default IconItalic