+++ react/IconUnorderedList.js
@@ -1,0 +1,3 @@
+import IconUnorderedList from './components/icon/UnorderedList/index'
+
+export default IconUnorderedList